+++ temporal-sdk/src/test/java/io/temporal/workflow/signalTests/SignalExternalWorkflowFailureTest.java
@@ -19,26 +19,39 @@
 
 package io.temporal.workflow.signalTests;
 
+import static io.temporal.api.enums.v1.SignalExternalWorkflowExecutionFailedCause.SIGNAL_EXTERNAL_WORKFLOW_EXECUTION_FAILED_CAUSE_EXTERNAL_WORKFLOW_EXECUTION_NOT_FOUND;
 import static org.junit.Assert.*;
 
 import io.temporal.api.common.v1.WorkflowExecution;
+import io.temporal.client.WorkflowClient;
 import io.temporal.client.WorkflowFailedException;
 import io.temporal.client.WorkflowOptions;
+import io.temporal.client.WorkflowStub;
 import io.temporal.failure.ApplicationFailure;
 import io.temporal.workflow.Workflow;
 import io.temporal.workflow.shared.SDKTestWorkflowRule;
 import io.temporal.workflow.shared.TestWorkflows.TestSignaledWorkflow;
 import io.temporal.workflow.shared.TestWorkflows.TestWorkflow1;
+import io.temporal.workflow.shared.TestWorkflows.TestWorkflowReturnString;
 import java.time.Duration;
+import java.util.concurrent.CountDownLatch;
 import org.junit.Rule;
 import org.junit.Test;
 
 public class SignalExternalWorkflowFailureTest {
 
+  private static String WORKFLOW_ID = "testWorkflowID";
+  private static String TEST_SIGNAL = "test";
+  private static CountDownLatch latch = new CountDownLatch(1);
+  private static CountDownLatch latch2 = new CountDownLatch(1);
+
   @Rule
   public SDKTestWorkflowRule testWorkflowRule =
       SDKTestWorkflowRule.newBuilder()
-          .setWorkflowTypes(TestSignalExternalWorkflowFailure.class)
+          .setWorkflowTypes(
+              TestSignalExternalWorkflowFailure.class,
+              WorkflowWithSignalImpl.class,
+              SignalingWorkflowImpl.class)
           .build();
 
   @Test
@@ -63,6 +76,34 @@ public class SignalExternalWorkflowFailureTest {
     }
   }
 
+  @Test
+  public void testTerminateWorkflowSignalError() throws InterruptedException {
+    WorkflowOptions options =
+        WorkflowOptions.newBuilder()
+            .setTaskQueue(testWorkflowRule.getTaskQueue())
+            .setWorkflowId(WORKFLOW_ID)
+            .build();
+    TestSignaledWorkflow terminatedWorkflow =
+        testWorkflowRule.getWorkflowClient().newWorkflowStub(TestSignaledWorkflow.class, options);
+    WorkflowClient.start(terminatedWorkflow::execute);
+
+    TestWorkflowReturnString signalingWorkflow =
+        testWorkflowRule.newWorkflowStub(TestWorkflowReturnString.class);
+    WorkflowClient.start(signalingWorkflow::execute);
+
+    // Wait for terminatedWorkflow to start
+    latch.await();
+
+    WorkflowStub stub = WorkflowStub.fromTyped(terminatedWorkflow);
+    stub.terminate("Mock terminating workflow");
+
+    // Wait for signalingWorkflow to start and terminatedWorkflow to terminate
+    latch2.countDown();
+
+    WorkflowStub workflowStub2 = WorkflowStub.fromTyped(signalingWorkflow);
+    assertEquals(workflowStub2.getResult(String.class), "Success!");
+  }
+
   public static class TestSignalExternalWorkflowFailure implements TestWorkflow1 {
 
     @Override
@@ -73,6 +114,49 @@ public class SignalExternalWorkflowFailureTest {
           Workflow.newExternalWorkflowStub(TestSignaledWorkflow.class, parentExecution);
       workflow.signal("World");
       return "ignored";
+    }
+  }
+
+  public static class WorkflowWithSignalImpl implements TestSignaledWorkflow {
+    private String signal;
+
+    @Override
+    public String execute() {
+      latch.countDown();
+      Workflow.await(Duration.ofDays(100500), () -> TEST_SIGNAL.equals(signal));
+      return signal;
+    }
+
+    @Override
+    public void signal(String signal) {
+      this.signal = signal;
+    }
+  }
+
+  public static class SignalingWorkflowImpl implements TestWorkflowReturnString {
+    @Override
+    public String execute() {
+      try {
+        latch2.await();
+        TestSignaledWorkflow stub =
+            Workflow.newExternalWorkflowStub(TestSignaledWorkflow.class, WORKFLOW_ID);
+        // Signaling a terminated workflow should not succeed
+        stub.signal(TEST_SIGNAL);
+        fail();
+      } catch (Exception e) {
+        if (e instanceof ApplicationFailure) {
+          // Expected
+          ApplicationFailure failure = (ApplicationFailure) e;
+          assertEquals(
+              failure.getType(),
+              SIGNAL_EXTERNAL_WORKFLOW_EXECUTION_FAILED_CAUSE_EXTERNAL_WORKFLOW_EXECUTION_NOT_FOUND
+                  .name());
+        } else {
+          // Something funky happened
+          fail();
+        }
+      }
+      return "Success!";
     }
   }
 }
